+++ components/CustomAssets/NativeAsset/NativeAsset.tsx
@@ -11,7 +11,7 @@ import { ColumnProps } from 'components/Table/types'
 import useAsset from 'hooks/useAsset'
 import { AssetType } from 'types'
 import { FishIcon } from 'svgx'
-import { ACTIONS_COLUMN } from 'components/Table/Table'
+import { ACTIONS_COLUMN } from 'components/ExplorerCommonTable'
 import RoutePaths from 'constants/RoutePaths'
 import { TableComponentProps } from '@ironfish/ui-kit/dist/components/Table/types'
 import { ASSET_COLUMNS } from '../CustomAssetsTable/CustomAssetsTable'
